Move component TargetPopulation to folder root
