+++ src/components/target-population/PopulationDistribution.tsx
@@ -6,8 +6,8 @@ import { createStyles, WithStyles, Theme } from "@material-ui/core";
 import { Table, TableRow, TableHead, TableCell, TableBody } from "@material-ui/core";
 import { withStyles } from "@material-ui/core/styles";
 
-import { memoize } from "../../../utils/memoize";
-import i18n from "../../../locales";
+import { memoize } from "../../utils/memoize";
+import i18n from "../../locales";
 import EditButton from "./EditButton";
 import { getValue, getShowValue } from "./utils";
 import Value from "./Value";
@@ -15,11 +15,11 @@ import {
     TargetPopulationItem,
     PopulationDistribution,
     TargetPopulation,
-} from "../../../models/TargetPopulation";
-import { OrganisationUnit, OrganisationUnitLevel, Maybe } from "../../../models/db.types";
+} from "../../models/TargetPopulation";
+import { OrganisationUnit, OrganisationUnitLevel, Maybe } from "../../models/db.types";
 import OrgUnitName from "./OrgUnitName";
 import "./PopulationDistribution.css";
-import { NumericField } from "../../shared/NumericField";
+import { NumericField } from "../shared/NumericField";
 
 export interface PopulationDistributionProps extends WithStyles<typeof styles> {
     organisationUnitLevels: OrganisationUnitLevel[];
